fixed a build error raised when compiling with C++14

+++ src/DDRTree.cpp
@@ -42,7 +42,7 @@ void pca_projection_cpp(const MatrixXd& C, int dimensions,  MatrixXd& W){
     std::vector<std::pair<double,MatrixXd::Index> > D;
     D.reserve(eVals.size());
     for (MatrixXd::Index i=0;i<eVals.size();i++)
-        D.push_back(std::make_pair<double,MatrixXd::Index>(eVals.coeff(i,0),i));
+        D.push_back(std::make_pair<double,MatrixXd::Index>((double)eVals.coeff(i,0),(long)i));
     std::sort(D.rbegin(),D.rend());
     MatrixXd sortedEigs;
     sortedEigs.resize(eVecs.rows(), dimensions);
